pyproject.toml
[project]
name = "pebble-productivity-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "gunicorn>=25.1.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
]
